+++ packages/facets/package-json/src/tasks/index.ts
@@ -1,5 +1,6 @@
 export { AddScript, PACKAGE_JSON__ADD_SCRIPT } from './add-script'
 export { RemoveScript, PACKAGE_JSON__REMOVE_SCRIPT } from './remove-script'
+export { RenameScript, PACKAGE_JSON__RENAME_SCRIPT } from './rename-script'
 export { ResolveConflict, PACKAGE_JSON__RESOLVE_CONFLICT } from './resolve-conflict'
 export { SetVersion, PACKAGE_JSON__SET_VERSION } from './set-version'
 export { SetProperty, PACKAGE_JSON__SET_PROPERTY } from './set-property'